+++ pefacil-product/pefacil-product-domain/src/main/java/br/com/pefacil/product/domain/port/api/FindProduct.java
@@ -7,7 +7,7 @@ import br.com.pefacil.product.domain.model.Product;
 
 public interface FindProduct {
 
-    Optional<Product> findById(Integer id);
+    Optional<Product> findById(Long id);
 
     Optional<Product> findByName(String name);
